Renamed date classes
