+++ sdk/src/main/java/cloud/mindbox/mobile_sdk/models/operation/DateTime.kt
@@ -1,8 +1,8 @@
-package cloud.mindbox.mobile_sdk.models.operation.request
+package cloud.mindbox.mobile_sdk.models.operation
 
 import java.util.*
 
-class DateOnlyRequest : Date {
+class DateTime : Date {
 
     constructor() : super()
