+++ sdk/src/main/java/cloud/mindbox/mobile_sdk/models/operation/adapters/DateOnlyAdapter.kt
@@ -1,6 +1,6 @@
 package cloud.mindbox.mobile_sdk.models.operation.adapters
 
-import cloud.mindbox.mobile_sdk.models.operation.request.DateOnlyRequest
+import cloud.mindbox.mobile_sdk.models.operation.DateOnly
 import cloud.mindbox.mobile_sdk.returnOnException
 import com.google.gson.TypeAdapter
 import com.google.gson.stream.JsonReader
@@ -9,11 +9,11 @@ import com.google.gson.stream.JsonWriter
 import java.text.SimpleDateFormat
 import java.util.*
 
-class DateOnlyRequestAdapter : TypeAdapter<DateOnlyRequest>() {
+class DateOnlyAdapter : TypeAdapter<DateOnly>() {
 
     private val formatter = SimpleDateFormat("yyyy-MM-dd", Locale.getDefault())
 
-    override fun write(out: JsonWriter?, value: DateOnlyRequest?) {
+    override fun write(out: JsonWriter?, value: DateOnly?) {
         runCatching {
             if (value == null) {
                 out?.nullValue()
@@ -23,14 +23,14 @@ class DateOnlyRequestAdapter : TypeAdapter<DateOnlyRequest>() {
         }.returnOnException { out }
     }
 
-    override fun read(`in`: JsonReader?): DateOnlyRequest? = `in`?.let { reader ->
+    override fun read(`in`: JsonReader?): DateOnly? = `in`?.let { reader ->
         runCatching {
             if (reader.peek() === JsonToken.NULL) {
                 reader.nextNull()
                 return@let null
             }
 
-            reader.nextString()?.let { formatter.parse(it)?.time?.let(::DateOnlyRequest) }
+            reader.nextString()?.let { formatter.parse(it)?.time?.let(::DateOnly) }
         }.returnOnException { null }
     }
