+++ sdk/src/main/java/cloud/mindbox/mobile_sdk/models/operation/request/CatalogProductListRequest.kt
@@ -1,12 +1,13 @@
 package cloud.mindbox.mobile_sdk.models.operation.request
 
-import cloud.mindbox.mobile_sdk.models.operation.adapters.DateTimeRequestAdapter
+import cloud.mindbox.mobile_sdk.models.operation.DateTime
+import cloud.mindbox.mobile_sdk.models.operation.adapters.DateTimeAdapter
 import com.google.gson.annotations.JsonAdapter
 import com.google.gson.annotations.SerializedName
 
 open class CatalogProductListRequest(
-    @JsonAdapter(DateTimeRequestAdapter::class)
-    @SerializedName("calculationDateTimeUtc") val calculationDateTimeUtc: DateTimeRequest? = null,
+    @JsonAdapter(DateTimeAdapter::class)
+    @SerializedName("calculationDateTimeUtc") val calculationDateTimeUtc: DateTime? = null,
     @SerializedName("area") val area: AreaRequest? = null,
     @SerializedName("items") val items: List<RequestedPromotionRequest>? = null
 )
